HMAN-32 - Code review comments
